Check if configuration use Absolute path not below Base directory
This shall not be to prevent "CHROOT defect".

- Configuration is checked
- Rule creation or modification through DAO will checked

Also includes fix and optimization of LongUuid

Fix issue #103

+++ WaarpCommon/src/main/java/org/waarp/common/guid/JvmProcessId.java
@@ -74,13 +74,13 @@ public final class JvmProcessId {
   }
 
   /**
-   * Use both PID and MAC address but as 16 bytes
+   * Use both PID and MAC address but as 8 bytes hash
    *
    * @return one id as much as possible unique
    */
-  public static short jvmInstanceId() {
+  public static byte jvmInstanceId() {
     final long id = 31 * jvmProcessId() + macAddressAsInt();
-    return (short) (Long.hashCode(id) % Short.MAX_VALUE);
+    return (byte) (Long.hashCode(id) % Byte.MAX_VALUE);
   }
 
   /**
